+++ flowmachine/tests/test_contact_modal_location_distance.py
@@ -5,6 +5,7 @@
 from flowmachine.features.subscriber.contact_modal_location_distance import *
 
 import pytest
+from flowmachine.features.subscriber.contact_balance import ContactBalance
 
 
 @pytest.mark.parametrize(
@@ -20,5 +21,16 @@ def test_contact_modal_location_distance(get_dataframe, statistic, msisdn, want)
     print(query.get_query())
     df = get_dataframe(query).set_index("subscriber")
     print(df.sample(n=5))
-    assert df.loc[msisdn, f"distance_{statistic}"] == pytest.approx(want)
+    assert df.value[msisdn] == pytest.approx(want)
 
+@pytest.mark.parametrize("kwarg", ["statistic", "contact_balance"])
+def test_contact_modal_location_errors(kwarg):
+    """ Test ValueError is raised for non-compliant kwarg in ContactModalLocationDistance. """
+
+    error = (
+        "error" if kwarg not in {"contact_balance"} 
+        else ContactBalance("2016-01-03", "2016-01-05", subscriber_identifier="imei")
+    )
+
+    with pytest.raises(ValueError):
+        query = ContactModalLocationDistance("2016-01-03", "2016-01-05", **{kwarg: error})